pyproject.toml
[tool.poetry]
name = "spaprodject"
version = "0.1.0"
description = ""
authors = ["Kostiantyn <lynx87@ukr.net>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
djangorestframework = "^3.14.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
